Use SegmentedBuffer in InspectorNetworkAgent
InspectorNetworkAgent's InspectorFileReaderLoaderClient is using a
SharedBuffer for appending data while reading data from a
BlobDataHandle. SharedBuffer is a refcounted SegmentedBuffer. But we
don't need to share the data in the InspectorFileReaderLoaderClient.
So to make the ownership clear, this CL changes
InspectorFileReaderLoaderClient to use a SegmentedBuffer to append the
data while reading data from a BlobDataHandle.
This CL also changes SharedBufferContent() method to accept
a pointer of a SegmentedBuffer instead of a SharedBuffer.

Bug: 342859647
Change-Id: I059709be452e390dc9ebd28dc058a09b3c9c6fda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5595330
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311259}