Use SegmentedBuffer for popup window and ForceSynchronousDocumentInstall
Currently LocalFrame::ForceSynchronousDocumentInstall() is accepting a
SharedBuffer. SharedBuffer is a refcounted SegmentedBuffer. But we don't
need to share the data after calling this method. So this CL changes
this method to accept a SegmentedBuffer.

This CL also changes PagePopupClient's methods to accept a
SegmentedBuffer instead of a SharedBuffer.

Bug: 342859647
Change-Id: I799d446848ca9afe636bd4a07a21ea82fd5d9397
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5595373
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311270}