Stop using SharedBuffer for indexeddb's key and value
Both IDBKey and IDBValue are using SharedBuffer for keeping their data.
But we don't need to use SharedBuffer that is a refcounted segmented
buffer, because their data must not be segmented.

IDBKey need a refcounted binary data. And IDBValue need a binary data.
So this CL makes IDBKey use a base::RefCountedData<Vector<char>> for its
data, and IDBValue use a Vector<char> for its data.

Bug: 342859647
Change-Id: I812d67d7f4f22ef4668e2f108391ac3d81263e46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5588216
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311275}