Use SegmentedBuffer in InspectorCacheStorageAgent
InspectorCacheStorageAgent CachedResponseFileReaderLoaderClient is using
a SharedBuffer for appending data while reading data from a
BlobDataHandle. SharedBuffer is a refcounted SegmentedBuffer. But we
don't need to share the data in the
CachedResponseFileReaderLoaderClient. So to make the ownership clear,
this CL changes CachedResponseFileReaderLoaderClient to use a
SegmentedBuffer to append the data while reading data from a
BlobDataHandle.

Bug: 342859647
Change-Id: I22b0890078a0e415ff0e2008897f7072e8a15afa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5595079
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311276}