[PWA] Remove FakeWebAppFileHandlerManager
This CL removes the FakeWebAppFileHandlerManager, and changes all
usages of it to use the real WebAppFileHandlerManager. The unit-tests
have been moved to the same test file as FileHandlingSubManager to
use existing Synchronization workflow to test the E2E logic, thus
reducing the tight coupling in the previous implementation.

This helps make the removal of the actual WebAppFileHandlerManager
a lot easier, and hence has been split into a separate CL. This
should technically be a no-op.

UPDATE: After 3-4 days of repititive debugging and running swarming
tests, the flakiness on Mac is still not easily reproducible. Since
these tests have merely been moved with little change in logic,
the `Validate-Test-Flakiness` flag is being used to bypass that check.
The flakiness will be being investigated separately by crbug.com/345367397.

Bug: 267330807, 1401125
Validate-Test-Flakiness: skip
Change-Id: Iaed6336a3b4d94eee1202cdb46bac8f1c5d38aaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5585526
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311287}