Use one of 7 sizes for Bidding and Auction Services API requests
Limit Bidding and Auction Services API encrypted requests to one of the
following 7 sizes: [0, 5, 10, 20, 30, 40, 55] kB. Requests will be
padded up to these lengths when a specific size is not specified. Note
that this also sets a maximum size for the encrypted request instead of
using arbitrarily large powers of 2.

This is necessary to limit the side channel where the size of the
encrypted blob can leak information about the presence of interest
groups on device. With 7 potential sizes, we limit the leak to less than
3 bits as a trade-off between privacy and utility. Previously we were
padding up to powers of 2 in size but had no max size so the encrypted
blob size could potentially leak relatively large amounts of
information, limited only by the maximum size for interest groups.

Change-Id: I05a5b405c7f46c6130974fdc56b829abc0908db2
Fixed: 343719161
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5577617
Reviewed-by: Caleb Raitto <caraitto@chromium.org>
Commit-Queue: Russ Hamilton <behamilton@google.com>
Cr-Commit-Position: refs/heads/main@{#1311307}