Always send gesture begin/end events for all taps on all platforms
We have seen some INP regression for the suppress toolbar feature,
which may be caused by the tap interactions on the webpages. We think
the INP regression of the feature of suppress toolbar capture may be
caused by the pointer events and click events generated from the tap
interaction. The latency of the click events includes the waiting time
of the acks of touch start/end events, which are blocked by the
processing of the capture.

Please see this doc for more details:
go/inp_regression_suppress_capture

In this CL, we are enabling gesture begin/end events to indicate the
the first and last gesture for a tap interaction on Android, which will
be used as a signal for starting capture in a following up CL.

Bug: 344965476, 265479149

Change-Id: I3b38a5de22a62e2dbce6ee06617efca96e9a0e17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5512503
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311311}