autotest: Fix stopSmoothnessTracking crash
The crash happens because `ForwardFrameRateDataAndReset` is
invoked when `stopSmoothnessTracking` times out. The report
callback will be invoked even though `ui::ThroughputTracker`
is destructed with its owner `DisplaySmoothnessTracker>`.

This CL adds a `ui::ThroughputTracker::CancelReport` is
cancel the reporting and call this API on timeouts.

Bug: b/328517929
Change-Id: I9f988a9bd6e52c981af73edf760512fd6885024b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5598482
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311315}