Reapply "webauthn: make attachment=any default to platform authenticators."
This reverts commit bbc6b9bd1a7cbca0043c7d2f456dc2b092ded6ce.

Timeline:
  #1310844: good build on both Mac11 and Mac13 tests
  #1310859: "make attachment=any default to platform authenticators"
  #1310882: bad build on both Mac11 and Mac13 tests
            (MakeCredentialDeclineGPM is bad)
  #1310997: MakeCredentialDeclineGPM disabled
  #1311007: revert "make attachment=any"...

It sure looks like the change broke this test. However, even after the
revert the test is still broken. After syncing to #1310844, the test is
_still_ broken when previously it worked.

I'm not sure what happened here, but this change didn't seem to break
it. From poking at things, the test depends on the state of iCloud
Drive, which I should fix. I'm guessing that some corp policy push
changed that and it happened at the same time as this change originally
landed.

This this change is important for M127, I'm relanding it without
changes. Then I'll follow up to reenable the test.

Bug: 343877496,345308672
Change-Id: I83dba8813e9e519dd1b7417ecf5b35e5bb356ab6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5605203
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311359}