kbd-backlight: Handle restore ALS on first login for new user.
This change follows the same pattern and aligns with the logic of
restoring display brightness behavior

With this change, the user's keyboard ALS settings will be saved. When
they log into a new device, the ALS will automatically be restored by
default on that device(true by default if unset).

Bug: b/326009607
Test: ash_unittests -f KeyboardBrightnessControllerTest
Change-Id: I1ee8dd98f2beb09bce99a14db279bbcfee1a1ae3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5593577
Commit-Queue: Longbo Wei <longbowei@google.com>
Reviewed-by: David Padlipsky <dpad@google.com>
Cr-Commit-Position: refs/heads/main@{#1311376}