[Tab Group Sync] Fix conflict resolution logic
This CL fixes the sync merge logic. After this CL and after
https://crrev.com/c/5597796, all the sync merge logic will live in the
MergeConflict function which is the recommended and less error-prone
way of handling updates.

With this, the conflict resolution logic is:
1. If both client and server have unmerged updates, the side with
last timestamp wins and overwrites the other side.
2. If the entity is deleted on server and updated on client, client
will be preferred.

The timestamp is based on the update timestamp set on the specifics
which is updated on any kind of update to the tab group, locally
generated via base::Time::Now().

Bug: 344798763
Change-Id: I30c63007a289ad177bd1fd6554ac1fb0b6ae89ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5594779
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/main@{#1311397}