[IOSSignout] Idle timeout policy string update for managed account
This CL updates the subtitle string for when IdleTimeoutActions is set
to sign out or sign out and close tabs. The new string is shown in the
dialog displayed when idle timeout happens while the user is signed in
to a managed account.

Screenshot: https://drive.google.com/file/d/1rgUnfztOlm2Pvg0FuxUvNZvwsNTNPLAD/view?usp=sharing

Low-Coverage-Reason: the coordinator is hard to test on its own, so it is covered by egtests
Bug: 340599079
Change-Id: I8cf14024bdaaa8faf85c155a3bde55fc4d34866e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5592992
Reviewed-by: Vincent Boisselle <vincb@google.com>
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Cr-Commit-Position: refs/heads/main@{#1311398}