[Read aloud] Improve screenReader access to language download messages
Currently, when users download a new language for read aloud, a visual
status message is shown, but this is not read by screen readers.

This CL changes the aria-live region value from "off" to "polite" so
that status messages are announced.

With ChromeVox on ChromeOS, the 'Downloading voices' message is
announced several times if the user takes no other action. This is
unnecessary, but an overall improvement for accessibility.


Bug: 40927698
Change-Id: I7417fcbdf69254228ed60f662945be104318340c
Fixed: 345272752
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5602195
Commit-Queue: Shari Trewin <trewin@google.com>
Reviewed-by: Eitan Goldberger <eitang@google.com>
Reviewed-by: Lauren Winston <lwinston@google.com>
Cr-Commit-Position: refs/heads/main@{#1311419}