Add internal tree test expectations for Android accessibility
Note to sheriffs: If a test begins to flake, please disable the test
                  directly rather than reverting, or contact mschillaci
                  to fix.

This CL adds many android internal tree dump tests. These are tests for
which there are already external tree tests, but for some unknown
reason there are no internal tests. We add the expectation files to
ensure that the internal tree is tested for all these cases.

AX-Relnotes: N/A
Bug: N/A
Change-Id: Ic11a615e81ec65080f026f6f30ad56ef78824b8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5599929
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1311424}