[SPC] Fix bug with no matching creds UX for inline network/issuers
The initial CL[0] that added the prototype inline network/issuer header
accidentally removed the logo for the no matching credential UX as well
as the transaction one. This was only for the case where the flag was
set, so there is no impact on the normal supported flow, but this CL
fixes the code under the flag to show the existing logo in the no
matching credentials case.

Longer term, we hope to experiment with a fallback UX concept that will
remove the no matching credentials UX entirely.

[0]: https://chromium-review.googlesource.com/c/chromium/src/+/5582451

Bug: 343801372
Change-Id: I471ba6d5655900ce423b6d3dcbaf07ed2bd94d14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5598618
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Auto-Submit: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311444}