Use per-subtype v8::CppHeapPointerTags for ScriptWrappables
v8 has recently introduced a scheme of specifying a tag when alling
v8::Object::Wrap and v8::Object::Unwrap to assert that pointers are
not having their types confused between wrapping and unwrapping.
Unwrapping takes a tag range rather than a single tag, in order to
support wrapping as a derived class but unwrapping as one of its
base classes.

This CL makes use of that scheme in the generated bindings for
ScriptWrappables. Types with inheritance are simply assigned a tag.
Inheritance trees are processed together, starting from the base
of the inheritance tree, and assigning tags in depth-first order.
Each class that acts as a base class therefore has a contiguous range
of all of its derived classes.

Bug: 342350850
Binary-Size: Regression is a tradeoff for a new security feature.
Fuchsia-Binary-Size: Regression is a tradeoff for a new security feature.
Change-Id: Ief0f230733f41cf3a008313348ea700588716136
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5581408
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311453}