ui/display: Make EdidParser retain edid_blob after construction
Currently, EdidParser is passed an const ref to edid_blob at
construction, which is used to parse. The edid blob (or reference to
it) is not retained after the constructor.

This CL changes the EdidParser so that it maintains a copy of the
edid_blob. The constructor is changed so that it accepts an rvalue
reference to the edid blob instead of a const ref, so that the blob is
moved and not copied. An accessor function is also added to return a
const ref to the edid blob, which is used in ozone/drm to eliminate a
redundant call to DRM to retrieve EDID.

Test: display_unittests, ozone_unittests, CQ for non-CrOS targets
Change-Id: I1c0a0537c0c04752df79f1c04e8a290794254e45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5564546
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Gil Dekel <gildekel@chromium.org>
Commit-Queue: Su Hong Koo <sukoo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311482}