Flags: Handle chrome://flags/deprecated in FlagsUI class.
Previously this was handled in a dedicated FlagsDeprecatedUI class,
which in turn is problematic because spawning a different
WebUIController subclass for chrome://flags vs chrome://flags/deprecated
is not possible when using the WebUIConfig mechanism.

Also adding a WebUI test that tests that chrome://flags/deprecated
works as expected.

Bug: 344643692
Change-Id: If710d715e0dc8f0802b5ee23af3ac85e39bb8abd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5594928
Commit-Queue: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: John Lee <johntlee@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1311509}