Merge ProductSpecifications & ProductSpecificationsSync flag
The sync integration for ProductSpecifications was developed under
a separate flag so the rest of the dev team/testers weren't blocked
in the event of unforeseen issues with the sync integration. Now
the sync integration is stable, the ProductSpecificationsSync flag is
being merged back with the ProductSpecifications flag.

Bug: b:345352628
Change-Id: I68975eadfe97cc38e1a81760d2a7ba237b5f91d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5598059
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: David Maunder <davidjm@chromium.org>
Reviewed-by: Ankush Singh <ankushkush@google.com>
Reviewed-by: Emily Shack <emshack@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311592}