webauthn: make security domain caching a Finchable fallback.
Caching the security domain results:
  * Goes wrong when the security domain is reset.
  * Doesn't work at all with a GPM PIN, where the code continues to
    believe that the security domain is empty after it just populated
    it.

The hope was that the non-GPM PIN launch would be out before needing to
worry about the second point, but that's not going to be true and also
the first point is obscure, but unpleasant.

Therefore disable caching, but keep it as a Finchable option if we find
that we're causing excess load.

Bug: 40274370
Change-Id: I35dff9a988a24c6b2fe956b248112f2c8a14a53b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5599078
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Auto-Submit: Adam Langley <agl@chromium.org>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311598}