[Reading mode] Make menu item selection status accessible
All of the menus in Reading mode use a checkmark to indicate the
currently selected item. This information was not available via a
screen reader. (This was also true in the original Views-based
toolbar).

This CL adds an aria-label to the checkmark, so it is announced
when it is visible.  I also considered using aria-current or
aria-selected but these were not appropriate for the menuitem role.
These options would also have read out the state after reading out
the item text and other role information. This approach provides the
information to users faster.

The associated string ends with a period so that when spoken by
a screenreader, the word "selected" is aurally separated from
the text of the option itself.

Bug: 40927698
Change-Id: Ie84c42a974499d820851b926617e37c3f4ed4632
Fixed: 345285432
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5605398
Commit-Queue: Shari Trewin <trewin@google.com>
Reviewed-by: Lauren Winston <lwinston@google.com>
Reviewed-by: Kristi Saney <kristislee@google.com>
Cr-Commit-Position: refs/heads/main@{#1311675}