Add Windows print backend load printable area test
To support testing, modify print_backend_win.cc so that it has a
separate header file.  Extend the class to include a testing callback
so tests can be notified whenever LoadPaperPrintableAreaUm() is called.

Demonstrate that the sequence of calling the PrintBackend APIs
GetPrinterSemanticCapsAndDefaults() followed by GetPaperPrintableArea()
for the same printer and paper ID result in two calls to
LoadPaperPrintableAreaUm().  This prepares for demonstrating an
upcoming optimization to avoid the second call when the same paper is
requested.

Bug: 40934068
Change-Id: I9885d40fdc97f9a3c9980ce4814a546bc9301b21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5606122
Commit-Queue: Alan Screen <awscreen@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311687}