Update GPU control list to use GPUCount method on gpu_info.
The gpu_count property in the driver_bug_list.json before this change
would count SW devices such as the Basic display renderer. When a user is remoted in, there can be extra Warp adapters in the adapter enumeration.

This means that the bug list behavior would change based on whether
chromium was started when the user was remoted in versus not.

This change uses the GpuCount accessor on gpu_info which eliminates
SW devices. It also fixes the entry for AV1 encoding which had a
workaround landed previously to address this issue

https://chromium-review.googlesource.com/c/chromium/src/+/3915479

Bug: 1367038
Change-Id: I16eb6869cb667f74ba5f10ef2318db804a7135db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5589790
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Maggie Chen <magchen@chromium.org>
Commit-Queue: Sushanth Rajasankar <Sushraja@microsoft.com>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Sushanth Rajasankar <Sushraja@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1311693}