[PM] Add NodeSetView
This new class provides a view into the nodes owned by the graph
without leaking implementation details.

This will allow the retirement of the VisitAll*Nodes() functions
in the graph that existed sorely for the purpose of avoiding full
container copies in favor of GetAll*Nodes() functions.

Change-Id: I46e0cfdab02a1b8c59b1a80b1908e0fb33b7322d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5592977
Reviewed-by: Joe Mason <joenotcharles@google.com>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1311712}